feat: remove bucket unencrypted
