+++ test/bucket.encription.s3-managed.test.ts
@@ -1,0 +1,39 @@
+import { App, Stack } from 'aws-cdk-lib';
+import { Match, Template } from 'aws-cdk-lib/assertions';
+import * as s3 from 'aws-cdk-lib/aws-s3';
+import { SecureBucket } from '../src';
+
+describe('SecureBucket Encryption Testing', () => {
+
+  const app = new App();
+  const stack = new Stack(app, 'TestingStack');
+  const bucket = new SecureBucket(stack, 'SecureBucket', {
+    bucketName: 'example-secure-bucket',
+    encryption: s3.BucketEncryption.S3_MANAGED,
+  });
+
+  it('Is Bucket', () => {
+    expect(bucket).toBeInstanceOf(s3.Bucket);
+  });
+
+  const template = Template.fromStack(stack);
+
+  it('Should have encryption sse algorithm aes256', () => {
+    template.hasResourceProperties('AWS::S3::Bucket', {
+      BucketEncryption: Match.objectEquals({
+        ServerSideEncryptionConfiguration: [
+          {
+            ServerSideEncryptionByDefault: {
+              SSEAlgorithm: 'AES256',
+            },
+          },
+        ],
+      }),
+    });
+  });
+
+  it('Should match snapshot', () => {
+    expect(template.toJSON()).toMatchSnapshot();
+  });
+
+});
